// src/data/models/templateDataModel.ts
export type blogTemplateMetaData = {
    templateIndex: number, // index of the template
    templateTitle: string, // title of the template
    templateDescription: string, // description of the template
    templateScreenShot: string, // screenshot of the template
    templateDemoLink: string, // link to the demo of the template
}